+++ Assets/Core/CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.4.1] - 2021-04-26 (Quest, Check for updates)
+## [1.5.0] - 2021-04-26 (Quest, Check for updates)
 ### Added
 * Add base class about [Quest] system.
 * Add check for updates menu editor class.
 * Add validate check for update Base frame package.
+* Add open data folder UnityEditor class. 
 
 ### Changed
 * Change the structure of changing from UGUI to ExtensionUI.